+++ ATK/Mock/FFTCheckerFilter.h
@@ -29,7 +29,7 @@ namespace ATK
     void set_checks(const std::vector<std::pair<int, DataType> >& frequency_checks);
     
   protected:
-    virtual void process_impl(int size);
+    virtual void process_impl(long size);
     virtual void setup();
     
     std::vector<std::pair<int, DataType> > frequency_checks;
