+++ ATK/Tools/SumFilter.cpp
@@ -22,7 +22,7 @@ namespace ATK
   }
 
   template<typename DataType_>
-  void SumFilter<DataType_>::process_impl(int size)
+  void SumFilter<DataType_>::process_impl(long size)
   {
     if(outputs_size[0] < size)
     {
@@ -30,7 +30,7 @@ namespace ATK
       outputs_size[0] = size;
     }
     
-    for(int i = 0; i < size; ++i)
+    for(long i = 0; i < size; ++i)
     {
       outputs[0][i] = static_cast<DataType>(converted_inputs[0][i] + converted_inputs[1][i]);
     }
